msgNotFound bug fixed & exception interface of autodeleting add

+++ bot/handlers/greetings.py
@@ -236,7 +236,7 @@ async def greeting_schedule_menu_cb(cb: CallbackQuery, callback_data: dict, stat
 @dp.callback_query_handler(greeting_action.filter(action="edit_send_delay"))
 async def edit_send_delay(cb: CallbackQuery, callback_data: dict, state: FSMContext):
     msg = await cb.message.answer(
-        "Введіть затримку надсилання (у секундах)",
+        "Введіть затримку надсилання у секундах (число від 1 до 86399)",
         reply_markup=gen_cancel(
             callback_data=greeting_action.new(
                 callback_data["id"],
@@ -252,7 +252,7 @@ async def edit_send_delay(cb: CallbackQuery, callback_data: dict, state: FSMCont
 @dp.message_handler(lambda msg: msg, content_types=ContentTypes.TEXT, state=states.InputStateGroup.greeting_send_delay)
 async def edit_send_delay(msg: Message, state: FSMContext):
     state_data = await state.get_data()
-    if msg.text.isdigit() and int(msg.text) > 0 and int(msg.text) < 86400:
+    if msg.text.isdigit() and 0 < int(msg.text) < 86400:
         await state.set_state(None)       
         greeting = greeting_db.get(state_data["greeting_id"])
         greeting.send_delay = int(msg.text)
@@ -263,16 +263,16 @@ async def edit_send_delay(msg: Message, state: FSMContext):
         await safe_del_msg(msg.from_user.id, msg.message_id)
         try:
             await bot.edit_message_text(
-                "❗️Затримка має бути цілим додатнім числом. Від 1 до 86399\nВведіть затримку надсилання (у секундах)",
+                "❗️Невірний формат\nВведіть затримку надсилання у секундах (число від 1 до 86399)",
                 msg.from_user.id,
                 state_data["msg_id"],
                 reply_markup=gen_cancel(
-                callback_data=greeting_action.new(
-                state_data["greeting_id"],
-                "schedule"  
+                    callback_data=greeting_action.new(
+                        state_data["greeting_id"],
+                        "schedule"
                     )
                 )
-                )
+            )
         except MessageNotModified:
             pass        
 
@@ -290,7 +290,7 @@ async def edit_del_delay(cb: CallbackQuery, callback_data: dict, state: FSMConte
     greeting = greeting_db.get(int(callback_data["id"]))
     bot_dc = bots_db.get(greeting.bot)
     msg = await cb.message.answer(
-        "Введіть затримку автовидалення (у секундах)",
+        "Введіть затримку надсилання у секундах (число від 1 до 86399)",
         reply_markup=gen_cancel(
             callback_data=greeting_action.new(
                 callback_data["id"],
@@ -303,15 +303,32 @@ async def edit_del_delay(cb: CallbackQuery, callback_data: dict, state: FSMConte
     await safe_del_msg(cb.from_user.id, cb.message.message_id)
 
 
-@dp.message_handler(lambda msg: msg.text.isdigit(), content_types=ContentTypes.TEXT, state=states.InputStateGroup.greeting_del_delay)
+@dp.message_handler(content_types=ContentTypes.TEXT, state=states.InputStateGroup.greeting_del_delay)
 async def edit_del_delay(msg: Message, state: FSMContext):
-    await state.set_state(None)
     state_data = await state.get_data()
-    greeting = greeting_db.get(state_data["greeting_id"])
-    greeting.del_delay = int(msg.text)
-    greeting_db.update(greeting)
-    await msg.delete()
-    await greeting_schedule_menu(msg.from_user.id, greeting.id, state_data["msg_id"])
+    if msg.text.isdigit() and 0 < int(msg.text) < 86400:
+        await state.set_state(None)
+        greeting = greeting_db.get(state_data["greeting_id"])
+        greeting.del_delay = int(msg.text)
+        greeting_db.update(greeting)
+        await msg.delete()
+        await greeting_schedule_menu(msg.from_user.id, greeting.id, state_data["msg_id"])
+    else:
+        await safe_del_msg(msg.from_user.id, msg.message_id)
+        try:
+            await bot.edit_message_text(
+                "❗️Невірний формат\nВведіть затримку надсилання у секундах (число від 1 до 86399)",
+                msg.from_user.id,
+                state_data["msg_id"],
+                reply_markup=gen_cancel(
+                    callback_data=greeting_action.new(
+                        state_data["greeting_id"],
+                        "schedule"
+                    )
+                )
+            )
+        except MessageNotModified:
+            pass
 
 
 @dp.callback_query_handler(greeting_action.filter(action="del_del_delay"))
